Function scatterplot + Admin option for sign in
Users (Admins) can now see their test results on the scatterplot:
- X axis is the test's user's ID
- Y axis is the test score
Additionally, there is a checkbox on the sign up page that allows the new user to choose whether or not they would like to be an admin. As of right now, the only difference between admins and regular users is that admins will see the scatterplot of every users' tests, while users will see a chart of only their own tests.